Rename `config` to `config.setupcfg`
This will facilitate the implementation of other configuration formats
(such as pyproject.toml as initially defined by PEP 621)
