+++ setuptools/config/setupcfg.py
