+++ setuptools/tests/config/test_setupcfg.py
@@ -9,9 +9,9 @@ import pytest
 from distutils.errors import DistutilsOptionError, DistutilsFileError
 from mock import patch
 from setuptools.dist import Distribution, _Distribution
-from setuptools.config import ConfigHandler, read_configuration
+from setuptools.config.setupcfg import ConfigHandler, read_configuration
 from distutils.core import Command
-from .textwrap import DALS
+from ..textwrap import DALS
 
 
 class ErrConfigHandler(ConfigHandler):
